Moved test plugin to separate package
