+++ force_bdss/core_plugins/test_data_sources/viscosity/viscosity_bundle.py
@@ -1,19 +1,20 @@
 from traits.api import provides, HasStrictTraits, String
 
-from force_bdss.data_sources.basic_data_source import BasicDataSource
 from force_bdss.data_sources.i_data_source_bundle import IDataSourceBundle
-from force_bdss.mco.basic_model import BasicModel
+
+from .viscosity_data_source import ViscosityDataSource
+from .viscosity_model import ViscosityModel
 
 
 @provides(IDataSourceBundle)
-class BasicBundle(HasStrictTraits):
-    name = String("basic")
+class ViscosityBundle(HasStrictTraits):
+    name = String("viscosity")
 
     def create_model(self, model_data):
-        return BasicModel.from_json(model_data)
+        return ViscosityModel.from_json(model_data)
 
     def create_ui(self):
         pass
 
     def create_data_source(self, application, model):
-        return BasicDataSource(self, application, model)
+        return ViscosityDataSource(self, application, model)
