+++ setup.py
@@ -12,7 +12,8 @@ setup(
         "force.bdss.extensions": [
             "mco = force_bdss.mco.multi_criteria_optimizers_plugin:"
             "MultiCriteriaOptimizersPlugin",
-            "data_source = force_bdss.data_sources.data_sources_plugin:"
+            "data_source = force_bdss.core_plugins."
+            "test_data_sources.data_sources_plugin:"
             "DataSourcesPlugin",
         ]
     },
